Moved all message-related types and macros to a new "message" module, renamed the message macro to "msg", updated its syntax a bit (using semicolons instead of commas to separate cases) and fixed the formatting.
